analysis: recover ANLZ cue colour from RGB (green, not orange)
A cue set on a CDJ stores palette index 0 with its real colour only in
the RGB bytes (the default hot cue is green, 0x00ff30). parsePCO2 read
the index byte alone, so external cues came back colour_id 0 and the
web painted them Pioneer orange. Read the RGB and, when the index is
0, recover the nearest palette id (0x00ff30 -> 0x16 green) — the same
rule the dbserver cue path already applied to CDJ-saved blobs.

To share that rule, the cue colour palette and NearestCueColorID move
from dbserver to analysis (which owns cue parsing and sits below
dbserver), and dbserver now delegates to them. Single source of truth
for the palette, no behaviour change on the dbserver side. Covered by
a test that a green-RGB PCO2 cue resolves to 0x16.
