+++ analysis/cuecolor.go
@@ -1,0 +1,46 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+package analysis
+
+// CueColorPalette maps a rekordbox hot-cue palette index to its RGB triplet,
+// mirroring the CUE_COLORS palette in the web UI (api/web/index.html). A deck
+// renders whatever RGB it is handed, so the only requirement is that the web
+// swatch and the on-deck colour stay consistent. Index 0x00 is Pioneer's
+// "no colour set" orange.
+var CueColorPalette = map[uint32][3]byte{
+	0x00: {0xff, 0x6a, 0x00},
+	0x01: {0x30, 0x5a, 0xff}, 0x02: {0x50, 0x73, 0xff}, 0x03: {0x50, 0x8c, 0xff}, 0x04: {0x50, 0xa0, 0xff},
+	0x05: {0x50, 0xb4, 0xff}, 0x06: {0x50, 0xb0, 0xf2}, 0x07: {0x50, 0xae, 0xe8}, 0x08: {0x45, 0xac, 0xdb},
+	0x09: {0x00, 0xe0, 0xff}, 0x0a: {0x19, 0xda, 0xf0}, 0x0b: {0x32, 0xd2, 0xe6}, 0x0c: {0x21, 0xb4, 0xb9},
+	0x0d: {0x20, 0xaa, 0xa0}, 0x0e: {0x1f, 0xa3, 0x92}, 0x0f: {0x19, 0xa0, 0x8c}, 0x10: {0x14, 0xa5, 0x84},
+	0x11: {0x14, 0xaa, 0x7d}, 0x12: {0x10, 0xb1, 0x76}, 0x13: {0x30, 0xd2, 0x6e}, 0x14: {0x37, 0xde, 0x5a},
+	0x15: {0x3c, 0xeb, 0x50}, 0x16: {0x28, 0xe2, 0x14}, 0x17: {0x7d, 0xc1, 0x3d}, 0x18: {0x8c, 0xc8, 0x32},
+	0x19: {0x9b, 0xd7, 0x23}, 0x1a: {0xa5, 0xe1, 0x16}, 0x1b: {0xa5, 0xdc, 0x0a}, 0x1c: {0xaa, 0xd2, 0x08},
+	0x1d: {0xb4, 0xc8, 0x05}, 0x1e: {0xb4, 0xbe, 0x04}, 0x1f: {0xba, 0xb4, 0x04}, 0x20: {0xc3, 0xaf, 0x04},
+	0x21: {0xe1, 0xaa, 0x00}, 0x22: {0xff, 0xa0, 0x00}, 0x23: {0xff, 0x96, 0x00}, 0x24: {0xff, 0x8c, 0x00},
+	0x25: {0xff, 0x75, 0x00}, 0x26: {0xe0, 0x64, 0x1b}, 0x27: {0xe0, 0x46, 0x1e}, 0x28: {0xe0, 0x30, 0x1e},
+	0x29: {0xe0, 0x28, 0x23}, 0x2a: {0xe6, 0x28, 0x28}, 0x2b: {0xff, 0x37, 0x6f}, 0x2c: {0xff, 0x2d, 0x6f},
+	0x2d: {0xff, 0x12, 0x7b}, 0x2e: {0xf5, 0x1e, 0x8c}, 0x2f: {0xeb, 0x2d, 0xa0}, 0x30: {0xe6, 0x37, 0xb4},
+	0x31: {0xde, 0x44, 0xcf}, 0x32: {0xde, 0x44, 0x8d}, 0x33: {0xe6, 0x30, 0xb4}, 0x34: {0xe6, 0x19, 0xdc},
+	0x35: {0xe6, 0x00, 0xff}, 0x36: {0xdc, 0x00, 0xff}, 0x37: {0xcc, 0x00, 0xff}, 0x38: {0xb4, 0x32, 0xff},
+	0x39: {0xb9, 0x3c, 0xff}, 0x3a: {0xc5, 0x42, 0xff}, 0x3b: {0xaa, 0x5a, 0xff}, 0x3c: {0xaa, 0x72, 0xff},
+	0x3d: {0x82, 0x72, 0xff}, 0x3e: {0x64, 0x73, 0xff},
+}
+
+// NearestCueColorID returns the palette color_id whose RGB is closest to the
+// given triplet (squared-Euclidean distance; ties resolve to the lower id so the
+// result is deterministic). Used when a cue carries a zero palette index but a
+// real colour in its RGB bytes — the CDJ encodes its default hot-cue green that
+// way, and taking index 0 alone would paint Pioneer orange (palette 0x00).
+func NearestCueColorID(r, g, b byte) uint32 {
+	best := uint32(0)
+	bestDist := -1
+	for idx, c := range CueColorPalette {
+		dr, dg, db := int(c[0])-int(r), int(c[1])-int(g), int(c[2])-int(b)
+		d := dr*dr + dg*dg + db*db
+		if bestDist < 0 || d < bestDist || (d == bestDist && idx < best) {
+			bestDist, best = d, idx
+		}
+	}
+	return best
+}
