+++ dbserver/cuepoints_test.go
@@ -67,20 +67,3 @@ func TestLoadAllRederivesCueColour(t *testing.T) {
 		t.Fatalf("re-derived colour = %+v, want one cue with color_id 0x16", cues)
 	}
 }
-
-func TestNearestCueColorID(t *testing.T) {
-	cases := []struct {
-		r, g, b byte
-		want    uint32
-		name    string
-	}{
-		{0x28, 0xe2, 0x14, 0x16, "exact palette green round-trips"},
-		{0x00, 0xff, 0x30, 0x16, "CDJ's brighter green maps to our green"},
-		{0xff, 0x6a, 0x00, 0x00, "orange maps back to idx 0"},
-	}
-	for _, c := range cases {
-		if got := nearestCueColorID(c.r, c.g, c.b); got != c.want {
-			t.Errorf("%s: nearest(%#x,%#x,%#x) = %#x, want %#x", c.name, c.r, c.g, c.b, got, c.want)
-		}
-	}
-}
